Replaced all existing references to NavBar/NavBarItem with new nav bar components
