--- components/ui/NavBar/_style.scss
@@ -1,19 +1,0 @@
-.navbar-drawer {
-  width: 300px;
-
-  &--header {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-
-    h6 {
-      padding-left: 15px;
-    }
-  }
-}
-
-.icon-btns-container {
-  margin-left: auto;
-  display: flex;
-}
