+++ components/ui/NavBarItem/index.tsx
@@ -1,56 +1,68 @@
-import React, { useState } from 'react';
-import {
-  Collapse, ListItem, ListItemText,
-} from '@material-ui/core';
+import React, { FunctionComponent, useState, useEffect } from 'react';
 import Link from 'next/link';
 import { Styled } from './_styles';
 
-interface Props {
-  title: string;
-  products: ProductItem[];
-}
-
-export type ProductItem = {
+type SubItem = {
   title: string;
   handle: string;
 };
 
-const NavBarItem = ({ title, products }: Props) => {
-  const [open, setOpen] = useState<boolean>(false);
+type Props = {
+  title: string;
+  handle: string;
+  products: SubItem[];
+};
 
-  const toggleOpen = () => {
-    setOpen(!open);
+const NavBarItem: FunctionComponent<Props> = ({ title, handle, products }: Props) => {
+  const [hover, setHover] = useState<boolean>(false);
+
+  const handleMouseOver = () => {
+    setHover(true);
+  };
+  const handleMouseLeave = () => {
+    setHover(false);
   };
 
-  const cleanHandle = (handle) => handle.substring(10, handle.length);
-
   return (
-    <>
-      <ListItem button onClick={toggleOpen} className="nav-list-item">
-        <ListItemText primary={title} />
-      </ListItem>
+    <Styled.Container
+      onFocus={handleMouseOver}
+      onMouseOver={handleMouseOver}
+      onMouseLeave={handleMouseLeave}
+      className={handle !== '#' ? handle : title.toLowerCase().replace(' ', '-')}
+    >
+      <Link
+        href={handle !== '#' ? { pathname: '/product', query: { handle } } : '#'}
+        as={handle !== '#' ? `/products/${handle}` : '#'}
+        passHref
+      >
+        <Styled.RootNavButton
+          hovered={hover}
+        >
+          <h4>{title}</h4>
+        </Styled.RootNavButton>
+      </Link>
       {products.length > 0 && (
-        <Collapse in={open} timeout="auto" unmountOnExit>
-          <Styled.NavList component="div" classes={{ root: 'nav-list-item--inner' }} disablePadding>
-            {products.map((product, index) => (
-              <Link
-                href={{
-                  pathname: '/product',
-                  query: { handle: cleanHandle(product.handle) },
-                }}
-                as={`/products/${cleanHandle(product.handle)}`}
-                passHref
-                key={index}
-              >
-                <ListItem button>
-                  <ListItemText primary={product.title} />
-                </ListItem>
-              </Link>
-            ))}
-          </Styled.NavList>
-        </Collapse>
+        <Styled.SubItemContainer
+          onFocus={handleMouseOver}
+          onMouseOver={handleMouseOver}
+          onMouseLeave={handleMouseLeave}
+          hovered={hover}
+        >
+          {products.map(({ title: subItemTitle, handle: subItemHandle }, key) => (
+            <Link
+              href={{ pathname: '/product', query: { handle: subItemHandle } }}
+              as={`/products/${subItemHandle}`}
+              passHref
+              key={key}
+            >
+              <Styled.SubItem>
+                <span>{subItemTitle}</span>
+              </Styled.SubItem>
+            </Link>
+          ))}
+        </Styled.SubItemContainer>
       )}
-    </>
+    </Styled.Container>
   );
 };
