--- components/ui/NewNavBar/navItems.json
@@ -1,43 +1,0 @@
-{
-  "navigationItems": [
-    {
-      "title": "Teeth Whitening",
-      "handle": "#",
-      "products": [
-        {
-          "title": "Teeth Whitening Device & Kit",
-          "handle": "teeth-whitening-led-device-kit"
-        }, {
-          "title": "Teeth Whitening Refill Serum",
-          "handle": "teeth-whitening-refill"
-        }
-      ]
-    }, {
-      "title": "Smart Toothbrushes",
-      "handle": "#",
-      "products": [
-        {
-          "title": "Smart Electric Toothbrush",
-          "handle": "smart-electric-toothbrush"
-        }, {
-          "title": "Smart Electric Toothbrush Refill Heads",
-          "handle": "smart-electric-toothbrush-refill"
-        }, {
-          "title": "Magik AR Toothbrush & App",
-          "handle": "magik-augmented-reality-toothbrush"
-        }, {
-          "title": "Smart Manual Toothbrush",
-          "handle": "m1"
-        }
-      ]
-    }, {
-      "title": "Get Connected",
-      "handle": "get-connected",
-      "products": []
-    }, {
-      "title": "Support",
-      "handle": "support",
-      "products": []
-    }
-  ]
-}
