+++ pages/cart.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Head from 'next/head';
 import { Typography, withTheme } from '@material-ui/core';
-import NewNavBar from 'components/ui/NewNavBar';
+import NavBar from 'components/ui/NavBar';
 import CartData from '../components/CartContent/CartData';
 import PageContainer, { PageSize } from '../views/layouts/PageContainer';
 import { withMuiApp } from '../hocs/withMui';
@@ -13,7 +13,7 @@ const CartPage = () => (
       <title>Products</title>
       <meta name="viewport" content="initial-scale=1.0, width=device-width" />
     </Head>
-    <NewNavBar />
+    <NavBar />
     <PageContainer paddingTop={30} size={PageSize.medium}>
       <Typography variant="h3" align="center">Cart</Typography>
       <CartData parentComponent="CartOverview" />
