+++ pages/login.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { ThemeProvider } from '@material-ui/styles';
 import { withTheme } from '@material-ui/core';
 import Head from 'next/head';
-import NewNavBar from 'components/ui/NewNavBar';
+import NavBar from 'components/ui/NavBar';
 import { theme } from '../views/theme';
 import { withMuiApp } from '../hocs/withMui';
 import LoginForm from '../components/LoginForm';
@@ -15,7 +15,7 @@ const LoginPage = () => (
       <title>Log In | ShopSmiles® by Colgate</title>
       <meta name="viewport" content="initial-scale=1.0, width=device-width" />
     </Head>
-    <NewNavBar />
+    <NavBar />
     <LoginForm />
     <FooterNav />
   </ThemeProvider>
